requirements.txt
opencv-python>=4.8.0
pandas>=2.0.0
openpyxl>=3.1.0
pyzbar>=0.1.9
qrcode[pil]>=7.4.0
Pillow>=10.0.0
